Move new endpoints to accounts controller instead of account controller
